+++ src/lib/constants.ts
@@ -1,4 +1,10 @@
-export const EMAIL_FORMAT_STORE_KEY = 'em_email_format'
-export const THEME_STORE_KEY = 'em_theme'
+export const STORE_KEYS = {
+  EMAIL_FORMAT: 'em_email_format',
+  AUTOFILL: 'em_autofill',
+  EXCLUDE_LIST: 'em_exclude_list',
+  EXCLUDE_LIST_PATTERNS: 'em_exclude_list_patterns',
+  THEME: 'em_theme'
+}
+
 export const DEFAULT_EMAIL_DOMAIN = 'maskmy.id'
 export const DEFAULT_EMAIL_FORMAT = `[domain][random:5]@example.${DEFAULT_EMAIL_DOMAIN}`